+++ lib/src/jsonize.dart
@@ -1,13 +1,33 @@
 import 'dart:convert';
 
+import 'helpers.dart';
+
 /// The encode and decode function prototype
-///
-/// The convertion function convetrs the serializable object instance into a
-/// string for [ConvertType.simple] convertion type, or any kind of jsonable
-/// type if convertion type is set to [ConvertType.complex].
 typedef ConvertFunction = dynamic Function(dynamic object);
 
-/// The [Jsonize] class used to trasform to and from json string
+/// The [DateTime] serialization format
+///
+/// - [string] a human readable date time string representation
+/// - [stringWithMillis] a human readable date time string representation
+///   with milliseconds
+/// - [stringWithMicros] a human readable date time string representation
+///   with microseconds
+/// - [epoch] a number representing the seconds since the "Unix epoch"
+///   1970-01-01T00:00:00Z (json space safer!)
+/// - [epochWithMillis] a number representing the milliseconds since the
+///   "Unix epoch" 1970-01-01T00:00:00Z
+/// - [epochWithMicros] a number representing the microseconds since the
+///   "Unix epoch" 1970-01-01T00:00:00Z
+enum DateTimeFormat {
+  string,
+  stringWithMillis,
+  stringWithMicros,
+  epoch,
+  epochWithMillis,
+  epochWithMicros
+}
+
+/// The [Jsonize] class used to transform to and from json string
 class Jsonize {
   /// The class token (you can change this token at runtime)
   ///
@@ -22,113 +42,78 @@ class Jsonize {
   /// ```
   static String jsonClassToken = "jt#";
 
-  /// Regiters a new type to the [Jsonize] convertion handling
+  /// Registers a new type to the [Jsonize] conversion handling
   static void registerType(Type classType, String classTypeCode,
       ConvertFunction? toJsonFunc, ConvertFunction? fromJsonFunc) {
-    // Some chechs on already registered types/classes
-    if (_encoders.containsKey(classType) &&
-        _encoders[classType]!.jsonClassCode != classTypeCode) {
-      throw "Class type '$classType' has already being registred with a different token! [${_encoders[classType]!.jsonClassCode} != $classTypeCode]";
+    // Some checks on already registered types/classes
+    if (encoders.containsKey(classType) &&
+        encoders[classType]!.jsonClassCode != classTypeCode) {
+      throw JsonizeException(
+          "Class type '$classType' has already being registred with a different"
+          " token! [${encoders[classType]!.jsonClassCode} != $classTypeCode]");
     }
-    if (_decoders.containsKey(classTypeCode) &&
-        _decoders[classTypeCode]!.classType != classType) {
-      throw "Class code '$classTypeCode' has already being registred with a different class! [${_decoders[classTypeCode]!.classType} != $classType]";
+    if (decoders.containsKey(classTypeCode) &&
+        decoders[classTypeCode]!.classType != classType) {
+      throw JsonizeException(
+          "Class code '$classTypeCode' has already being registred with a"
+          " different class! [${decoders[classTypeCode]!.classType}"
+          " != $classType]");
     }
-    _encoders[classType] = _ConvertInfo(classType, classTypeCode, toJsonFunc);
-    _decoders[classTypeCode] =
-        _ConvertInfo(classType, classTypeCode, fromJsonFunc);
+    encoders[classType] = ConvertInfo(classType, classTypeCode, toJsonFunc);
+    decoders[classTypeCode] =
+        ConvertInfo(classType, classTypeCode, fromJsonFunc);
   }
 
-  /// Regiters a new [Jsonizable] class by it instance.
+  /// Registers a new [Jsonizable] class by it instance.
   static void registerClass(Jsonizable object) {
     registerType(
         object.runtimeType, object.jsonClassCode, null, object.fromJson);
   }
 
-  /// Regiters new [Jsonizable] classes by it instances.
+  /// Registers new [Jsonizable] classes by it instances.
   static void registerClasses(Iterable<Jsonizable> objects) {
     for (var object in objects) {
       registerClass(object);
     }
   }
 
-  /// The [toJson] function transforming an object into a json string.
-  static dynamic toJson(dynamic value, [String? indent]) {
+  /// The [toJson] function transforms an object into a json string.
+  static String toJson(dynamic value,
+      {String? indent,
+      String? jsonClassToken,
+      String? dtClassCode,
+      DateTimeFormat? dateTimeFormat}) {
+    // Create a new session with requested parameters
+    JsonizeSession session = JsonizeSession(
+        jsonClassToken: jsonClassToken,
+        dtClassCode: dtClassCode,
+        dateTimeFormat: dateTimeFormat);
+    // Encode with the current session settings
     JsonEncoder encoder = indent == null
-        ? JsonEncoder(Jsonize._toEncodable)
-        : JsonEncoder.withIndent(indent, Jsonize._toEncodable);
+        ? JsonEncoder(session.toEncodable)
+        : JsonEncoder.withIndent(indent, session.toEncodable);
     return encoder.convert(value);
   }
 
-  /// The [fromJson] function transforming a json string back to an object.
-  static dynamic fromJson(dynamic value) {
-    return jsonDecode(value, reviver: _reviver);
+  /// The [fromJson] function transforms a json string back to an object.
+  static dynamic fromJson(dynamic value,
+      {String? jsonClassToken,
+      String? dtClassCode,
+      DateTimeFormat? dateTimeFormat}) {
+    // Create a new session with requested parameters
+    JsonizeSession session = JsonizeSession(
+        jsonClassToken: jsonClassToken,
+        dtClassCode: dtClassCode,
+        dateTimeFormat: dateTimeFormat);
+    // Decode with the current session settings
+    return jsonDecode(value, reviver: session.reviver);
   }
 
   /// The encode functions map
-  static final Map<Type, _ConvertInfo> _encoders = {
-    DateTime: _ConvertInfo(
-        DateTime, DateTimeJsonable.jsonClassCode, DateTimeJsonable.toJson),
-  };
+  static final Map<Type, ConvertInfo> encoders = {};
 
   /// The decode functions map
-  static final Map<String, _ConvertInfo> _decoders = {
-    DateTimeJsonable.jsonClassCode: _ConvertInfo(
-        DateTime, DateTimeJsonable.jsonClassCode, DateTimeJsonable.fromJson),
-  };
-
-  /// A convert helper function
-  static dynamic _convert(_ConvertInfo info, dynamic object) {
-    if (object is Jsonizable) {
-      return object.toJson();
-    }
-    return info.convert(object);
-  }
-
-  /// Makes a class token string
-  static String _makeClassToken(String classCode) {
-    return "$jsonClassToken$classCode";
-  }
-
-  /// Gets the class token string if the value reppresents a class token
-  static String? _getClassType(dynamic value) {
-    if (value is String && value.startsWith(jsonClassToken)) {
-      return value.substring(jsonClassToken.length);
-    }
-    return null;
-  }
-
-  /// The [_toEncodable] function used to convert to json string
-  static dynamic _toEncodable(dynamic object) {
-    // Find the class encoder
-    var convertInfo = _encoders[object.runtimeType];
-    if (convertInfo != null) {
-      return {
-        _makeClassToken(convertInfo.jsonClassCode):
-            _convert(convertInfo, object)
-      };
-    }
-    return object;
-  }
-
-  /// The [_reviver] function used to convert from json string
-  static dynamic _reviver(key, value) {
-    // Not the final value? (i.e. key == null)
-    if (key != null) {
-      var classType = _getClassType(key);
-      if (classType != null) {
-        var convertInfo = _decoders[classType];
-        if (convertInfo != null) {
-          return convertInfo.convert(value);
-        }
-      }
-    }
-    // Complex value? (i.e. incapsulated into the type map)
-    if (value is Map && value.keys.first.startsWith(jsonClassToken)) {
-      return value.values.first;
-    }
-    return value;
-  }
+  static final Map<String, ConvertInfo> decoders = {};
 }
 
 /// The interface to create a [Jsonizable] class you can register
@@ -151,24 +136,16 @@ class Jsonize {
 ///
 /// Jsonize.registerClass(MyClass.empty());
 /// ```
-
 abstract class Jsonizable<T> {
   String get jsonClassCode;
   dynamic toJson();
   T? fromJson(dynamic value);
 }
 
-/// The class used to register [DateTime] type serialization
-class DateTimeJsonable {
-  static String get jsonClassCode => "dt";
-  static dynamic toJson(DateTime v) => v.toIso8601String();
-  static DateTime? fromJson(dynamic v) => DateTime.tryParse(v.toString());
-}
-
-/// Internal class to keep convertion functions
-class _ConvertInfo {
-  final Type classType;
-  final String jsonClassCode;
-  final dynamic convert;
-  _ConvertInfo(this.classType, this.jsonClassCode, this.convert);
+/// The [Jsonize] package exception class
+class JsonizeException implements Exception {
+  final String msg;
+  const JsonizeException(this.msg);
+  @override
+  String toString() => '[Jsonize] $msg';
 }